# Optimizations for activating spans and noop spans