--- a/dd-trace-ot/src/main/java/datadog/opentracing/CustomScopeManagerWrapper.java
+++ b/dd-trace-ot/src/main/java/datadog/opentracing/CustomScopeManagerWrapper.java
@@ -3,6 +3,7 @@ package datadog.opentracing;
 import datadog.trace.bootstrap.instrumentation.api.AgentScope;
 import datadog.trace.bootstrap.instrumentation.api.AgentScopeManager;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
+import datadog.trace.bootstrap.instrumentation.api.AgentTracer;
 import datadog.trace.bootstrap.instrumentation.api.ScopeSource;
 import datadog.trace.context.TraceScope;
 import io.opentracing.Scope;
@@ -142,5 +143,10 @@ class CustomScopeManagerWrapper implements AgentScopeManager {
     public int hashCode() {
       return Objects.hash(delegate);
     }
+  }
+
+  @Override
+  public AgentScope activateNoopScope(ScopeSource source) {
+    return activate(AgentTracer.NoopAgentSpan.INSTANCE, source);
   }
 }