Add a special case for creating and activating a noop scope, that has a unique noop span, so it can be used to _clear_ the scope stack up to a specific point. 

Here are some benchmark results. Please note that the `dualNoopScope` case will be faster since it will short circuit and only increment a counter, as well as not call the scope listeners for the second scope.

```
Baseline
--------
Benchmark                                                             Mode  Cnt         Score         Error   Units
ActivateSpan.dualActivateNoopScope                                   thrpt    5  18829190.268 ±  253216.602   ops/s
ActivateSpan.dualActivateNoopScope:·gc.alloc.rate.norm               thrpt    5        48.000 ±       0.001    B/op
ActivateSpan.dualNoopScope                                           thrpt    5  21652267.455 ±  282995.940   ops/s
ActivateSpan.dualNoopScope:·gc.alloc.rate.norm                       thrpt    5        56.000 ±       0.001    B/op
ActivateSpan.dualScope                                               thrpt    5  14622150.909 ±  151840.159   ops/s
ActivateSpan.dualScope:·gc.alloc.rate.norm                           thrpt    5        64.000 ±       0.001    B/op
ActivateSpan.singleActivateNoopScope                                 thrpt    5  40515772.995 ±  141214.736   ops/s
ActivateSpan.singleActivateNoopScope:·gc.alloc.rate.norm             thrpt    5        ≈ 10⁻⁶                  B/op
ActivateSpan.singleNoopScope                                         thrpt    5  35828640.968 ± 2786135.485   ops/s
ActivateSpan.singleNoopScope:·gc.alloc.rate.norm                     thrpt    5        32.000 ±       0.001    B/op
ActivateSpan.singleScope                                             thrpt    5  33812503.141 ±  921420.259   ops/s
ActivateSpan.singleScope:·gc.alloc.rate.norm                         thrpt    5        32.000 ±       0.001    B/op

After optimizations
-------------------
Benchmark                                                             Mode  Cnt         Score         Error   Units
ActivateSpan.dualActivateNoopScope                                   thrpt    5  25208889.394 ±  157284.648   ops/s
ActivateSpan.dualActivateNoopScope:·gc.alloc.rate.norm               thrpt    5        ≈ 10⁻⁶                  B/op
ActivateSpan.dualNoopScope                                           thrpt    5  32535700.161 ± 1932828.890   ops/s
ActivateSpan.dualNoopScope:·gc.alloc.rate.norm                       thrpt    5        32.000 ±       0.001    B/op
ActivateSpan.dualScope                                               thrpt    5  21630399.414 ±  110936.129   ops/s
ActivateSpan.dualScope:·gc.alloc.rate.norm                           thrpt    5        64.000 ±       0.001    B/op
ActivateSpan.singleActivateNoopScope                                 thrpt    5  53024539.589 ±  233989.977   ops/s
ActivateSpan.singleActivateNoopScope:·gc.alloc.rate.norm             thrpt    5        ≈ 10⁻⁶                  B/op
ActivateSpan.singleNoopScope                                         thrpt    5  45480336.216 ±  940120.777   ops/s
ActivateSpan.singleNoopScope:·gc.alloc.rate.norm                     thrpt    5        32.000 ±       0.001    B/op
ActivateSpan.singleScope                                             thrpt    5  44521703.601 ± 3362219.385   ops/s
ActivateSpan.singleScope:·gc.alloc.rate.norm                         thrpt    5        32.000 ±       0.001    B/op
```