@richardstartin 2021-05-04
So this replaces the deque with an implicit linked list - can we run larger scale benchmarks for this, perhaps with akka where scope stacks get a lot deeper and may need to be unwound occasionally?