@bantonsson 2021-05-05
The numbers in the test have way too much noise in them even after adding _CPU_ work. The original motivation for the special casing of the noop scope creation has been made redundant by #2688 so I'll close this for now.